src/main/webapp/app/entities/client/route/client-routing-resolve.service.ts
import { inject } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of, EMPTY, Observable } from 'rxjs';
import { mergeMap } from 'rxjs/operators';

import { IClient } from '../client.model';
import { ClientService } from '../service/client.service';

export const clientResolve = (route: ActivatedRouteSnapshot): Observable<null | IClient> => {
  const id = route.params['id'];
  if (id) {
    return inject(ClientService)
      .find(id)
      .pipe(
        mergeMap((client: HttpResponse<IClient>) => {
          if (client.body) {
            return of(client.body);
          } else {
            inject(Router).navigate(['404']);
            return EMPTY;
          }
        })
      );
  }
  return of(null);
};

export default clientResolve;
